+++ packages/app/src/app/components/Create/utils/constants.ts
@@ -4,3 +4,15 @@ export const DEVBOX_BUTTON_DESCRIPTION =
   'Build and share standalone projects of any size in our Cloud Development Environment.';
 export const IMPORT_BUTTON_DESCRIPTION =
   'Run any branch instantly, create and review PRs in our Cloud Development Environment.';
+
+export const FEATURED_IDS = [
+  '9qputt', // react (vite + ts)
+  '3l5fg9', // javascript-devbox
+  'kmwy42', // html-css-devbox
+  'fxis37', // next
+  'k5hn76', // next.js + postgres
+  'pb6sit', // vue
+  'angular', // angular
+  'hsd8ke', // docker
+  'in2qez', // python
+];
